h5mgm/makefile: Change OBJ definition to point to ODIR instead of SDIR
The object compilation rule assumes the target is in ODIR.  But this
target was never being requested, as the h5mgm target requests the
object file from OBJ, which was in SDIR.  This means that an implicit
compilation rule was being used, rather than the explicit rule in the
makefile.  This led to very hard to debug problems with the wrong
compiler being used (confusion between CC from the environment, mpicc
set at the top, etc).